Address review: shorten playground buttons and use modals for load/save
- Shorten top-bar buttons to "Load prompt" / "Save prompt".
- Convert the load picker and the save dialog from side Drawers to
  Modals, for consistency with the "Create prompt" modal in the Prompt
  Registry page (rename SavePromptVersionDrawer -> SavePromptVersionModal).
